requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
openai==0.28
python-dotenv==1.0.1
playwright==1.42.0 